Move CLI test helpers to test utils
