+++ tests/unit/test_cli.py
@@ -1,25 +1,7 @@
 import os
 from imp import reload
-from subprocess import PIPE, Popen
 
-from .util import DjangoSetupMixin
-
-
-def run_silently(command):
-    """Run a shell command and return both exit_status and console output."""
-    command_args = command.split()
-    process = Popen(command_args, stdout=PIPE, stderr=PIPE, stdin=PIPE)
-    stdout, stderr = process.communicate()
-    output = (stdout.decode('UTF-8') + os.linesep +
-              stderr.decode('UTF-8')).strip() + os.linesep
-    return process.returncode, output
-
-
-def show_run_error(exit_status, output):
-    """An easy-to-read error message for assert"""
-    return 'Failed with exit status %s\n' \
-           '--------------\n' \
-           '%s' % (exit_status, output)
+from .util import DjangoSetupMixin, run_silently, show_run_error
 
 
 class TestCommandLine(DjangoSetupMixin):
